requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
requests
